Split out Customer into separate class.
Based on debate, decided to split up Customer class into separate
class. Made some tweaks to the guide as well.
